event actions [nfc]: Move registerAndStartPolling here, from fetchActions
Along with its helpers:

- registerStart
- registerAbortPlain
- registerAbort
- registerComplete
- registerAndStartPolling

(fetchPrivateMessages is also a helper, but would be kind of
annoying to uproot, including exporting messageFetchComplete, and it
has a TODO(server-2.1) for deletion anyway.)

This seems like a better place for these than something with
"message" in the path -- the events system deals with a lot more
than just messages.
