+++ src/message/fetchActions.js
@@ -1,6 +1,5 @@
 /* @flow strict-local */
 import * as logging from '../utils/logging';
-import * as NavigationService from '../nav/NavigationService';
 import type {
   Auth,
   Narrow,
@@ -11,13 +10,8 @@ import type {
   ThunkAction,
   UserId,
 } from '../types';
-import { ensureUnreachable } from '../types';
-import type { RegisterAbortReason } from '../actionTypes';
-import type { InitialData } from '../api/initialDataTypes';
 import * as api from '../api';
-import { ApiError, Server5xxError, NetworkError } from '../api/apiErrors';
-import { resetToAccountPicker } from '../nav/navActions';
-import { deadQueue } from '../session/sessionActions';
+import { Server5xxError, NetworkError } from '../api/apiErrors';
 import {
   getAuth,
   getSession,
@@ -25,33 +19,20 @@ import {
   getLastMessageId,
   getCaughtUpForNarrow,
   getFetchingForNarrow,
-  getIdentity,
   getStreamsById,
   getZulipFeatureLevel,
 } from '../selectors';
 import config from '../config';
 import {
-  REGISTER_START,
-  REGISTER_ABORT,
-  REGISTER_COMPLETE,
   MESSAGE_FETCH_START,
   MESSAGE_FETCH_ERROR,
   MESSAGE_FETCH_COMPLETE,
 } from '../actionConstants';
 import { FIRST_UNREAD_ANCHOR, LAST_MESSAGE_ANCHOR } from '../anchor';
-import { showErrorAlert } from '../utils/info';
 import { ALL_PRIVATE_NARROW, apiNarrowOfNarrow, caseNarrow, topicNarrow } from '../utils/narrow';
 import { BackoffMachine, promiseTimeout, TimeoutError } from '../utils/async';
 import { addToOutbox } from '../outbox/outboxActions';
-import { startEventPolling } from '../events/eventActions';
-import { logout } from '../account/logoutActions';
-import { ZulipVersion } from '../utils/zulipVersion';
 import { getAllUsersById, getOwnUserId } from '../users/userSelectors';
-import { getHaveServerData } from '../haveServerDataSelectors';
-import { MIN_RECENTPMS_SERVER_VERSION } from '../pm-conversations/pmConversationsModel';
-import { kNextMinSupportedVersion } from '../common/ServerCompatBanner';
-import { getOwnUserRole, roleIsAtLeast } from '../permissionSelectors';
-import { Role } from '../api/permissionsTypes';
 
 const messageFetchStart = (
   narrow: Narrow,
@@ -248,71 +229,6 @@ export async function fetchSomeMessageIdForConversation(
   return message?.id;
 }
 
-const registerStart = (): PerAccountAction => ({
-  type: REGISTER_START,
-});
-
-const registerAbortPlain = (reason: RegisterAbortReason): PerAccountAction => ({
-  type: REGISTER_ABORT,
-  reason,
-});
-
-export const registerAbort =
-  (reason: RegisterAbortReason): ThunkAction<Promise<void>> =>
-  async (dispatch, getState) => {
-    dispatch(registerAbortPlain(reason));
-    if (getHaveServerData(getState()) && reason !== 'unexpected') {
-      // Try again, forever if necessary; the user has an interactable UI and
-      // can look at stale data while waiting.
-      //
-      // Do so by lying that the server has told us our queue is invalid and
-      // we need a new one. Note that this must fire *after*
-      // `registerAbortPlain()`, so that AppDataFetcher sees
-      // `needsInitialFetch` go from `false` to `true`. We don't call
-      // `registerAndStartPolling` directly here because that would go against
-      // `AppDataFetcher`'s implicit interface.
-      //
-      // TODO: Clean up all this brittle logic.
-      // TODO: Instead, let the retry be on-demand, with a banner.
-      dispatch(deadQueue());
-    } else {
-      // Tell the user we've given up and let them try the same account or a
-      // different account from the account picker.
-      showErrorAlert(
-        // TODO: Set up these user-facing strings for translation once
-        // `initialFetchAbort`'s callers all have access to a `GetText`
-        // function. As of adding the strings, the initial fetch is dispatched
-        // from `AppDataFetcher` which isn't a descendant of
-        // `TranslationProvider`.
-        'Connection failed',
-        (() => {
-          const realmStr = getIdentity(getState()).realm.toString();
-          switch (reason) {
-            case 'server':
-              return roleIsAtLeast(getOwnUserRole(getState()), Role.Admin)
-                ? `Could not connect to ${realmStr} because the server encountered an error. Please check the server logs.`
-                : `Could not connect to ${realmStr} because the server encountered an error. Please ask an admin to check the server logs.`;
-            case 'network':
-              return `The network request to ${realmStr} failed.`;
-            case 'timeout':
-              return `Gave up trying to connect to ${realmStr} after waiting too long.`;
-            case 'unexpected':
-              return `Unexpected error while trying to connect to ${realmStr}.`;
-            default:
-              ensureUnreachable(reason);
-              return '';
-          }
-        })(),
-      );
-      NavigationService.dispatch(resetToAccountPicker());
-    }
-  };
-
-const registerComplete = (data: InitialData): PerAccountAction => ({
-  type: REGISTER_COMPLETE,
-  data,
-});
-
 /** Private; exported only for tests. */
 export const isFetchNeededAtAnchor = (
   state: PerAccountState,
@@ -468,109 +384,6 @@ export async function tryFetch<T>(
     throw e;
   }
 }
-
-/**
- * Connect to the Zulip event system for real-time updates.
- *
- * For background on the Zulip event system and how we use it, see docs from
- * the client-side perspective:
- *   https://github.com/zulip/zulip-mobile/blob/main/docs/architecture/realtime.md
- * and a mainly server-side perspective:
- *   https://zulip.readthedocs.io/en/latest/subsystems/events-system.html
- *
- * First does POST /register, which is sometimes called the "initial fetch".
- * because the response comes with a large payload of current-state data
- * that we fetch as part of initializing the event queue:
- *   https://zulip.readthedocs.io/en/latest/subsystems/events-system.html#the-initial-data-fetch
- *
- * Then immediately starts an async loop to poll for events.
- *
- * We fetch private messages here so that we can show something useful in
- * the PM conversations screen, but we hope to stop doing this soon (see
- * note at `fetchPrivateMessages`). We fetch messages in a few other places:
- * to initially populate a message list (`ChatScreen`), to grab more
- * messages on scrolling to the top or bottom of the message list
- * (`fetchOlder` and `fetchNewer`), and to grab search results
- * (`SearchMessagesScreen`).
- */
-export const registerAndStartPolling =
-  (): ThunkAction<Promise<void>> => async (dispatch, getState) => {
-    const auth = getAuth(getState());
-
-    const haveServerData = getHaveServerData(getState());
-
-    dispatch(registerStart());
-    let initData: InitialData;
-    try {
-      initData = await tryFetch(
-        // Currently, no input we're giving `registerForEvents` is
-        // conditional on the server version / feature level. If we
-        // need to do that, make sure that data is up-to-date -- we've
-        // been using this `registerForEvents` call to update the
-        // feature level in Redux, which means the value in Redux will
-        // be from the *last* time it was run. That could be a long
-        // time ago, like from the previous app startup.
-        () => api.registerForEvents(auth),
-        // We might have (potentially stale) server data already. If
-        // we do, we'll be showing some UI that lets the user see that
-        // data. If we don't, we'll be showing a full-screen loading
-        // indicator that prevents the user from doing anything useful
-        // -- if that's the case, don't bother retrying on 5xx errors,
-        // to save the user's time and patience. They can retry
-        // manually if they want.
-        haveServerData,
-      );
-    } catch (errorIllTyped) {
-      const e: mixed = errorIllTyped; // https://github.com/facebook/flow/issues/2470
-      if (e instanceof ApiError) {
-        // This should only happen when `auth` is no longer valid. No
-        // use retrying; just log out.
-        dispatch(logout());
-      } else if (e instanceof Server5xxError) {
-        dispatch(registerAbort('server'));
-      } else if (e instanceof NetworkError) {
-        dispatch(registerAbort('network'));
-      } else if (e instanceof TimeoutError) {
-        // We always want to abort if we've kept the user waiting an
-        // unreasonably long time.
-        dispatch(registerAbort('timeout'));
-      } else {
-        dispatch(registerAbort('unexpected'));
-        // $FlowFixMe[incompatible-cast]: assuming caught exception was Error
-        logging.warn((e: Error), {
-          message: 'Unexpected error during /register.',
-        });
-      }
-      return;
-    }
-
-    const serverVersion = new ZulipVersion(initData.zulip_version);
-
-    // Set Sentry tags for the server version immediately, so they're accurate
-    // in case we hit an exception in reducers on `registerComplete` below.
-    logging.setTagsFromServerVersion(serverVersion);
-
-    if (!serverVersion.isAtLeast(kNextMinSupportedVersion)) {
-      // The server version is either one we already don't support, or one
-      // we'll stop supporting the next time we increase our minimum supported
-      // version.  Warn so that we have an idea of how widespread this is.
-      // (Include the coarse version in the warning message, so that it splits
-      // into separate Sentry "issues" by coarse version.  The Sentry events
-      // will also have the detailed version, from the call above to
-      // `logging.setTagsFromServerVersion`.)
-      logging.warn(`Old server version: ${serverVersion.classify().coarse}`);
-    }
-
-    dispatch(registerComplete(initData));
-
-    dispatch(startEventPolling(initData.queue_id, initData.last_event_id));
-
-    // This is part of "register" in that it fills in some missing /register
-    // functionality for older servers; see fetchPrivateMessages for detail.
-    if (!serverVersion.isAtLeast(MIN_RECENTPMS_SERVER_VERSION)) {
-      dispatch(fetchPrivateMessages());
-    }
-  };
 
 export const uploadFile =
   (destinationNarrow: Narrow, uri: string, name: string): ThunkAction<Promise<void>> =>
